project1_data_collection/implementation/requirements.txt
pyfaktory==0.2.6
pymongo==4.10.1
requests==2.32.3
ratelimit==2.2.1
python-dotenv==1.0.1